requirements.txt
BeautifulSoup4
cjklib3
gevent
pandas
pypinyin
requests
